requirements.txt
asgiref==3.3.4
django-rest-authtoken==2.1.3
Django==3.2.4
djangorestframework==3.12.4
pip==21.1.2
pytz==2021.1
setuptools==57.0.0
sqlparse==0.4.1
